add lib.rs
